backend/requirements.txt
# Core
biopython>=1.84
langchain>=0.3.0
langchain-ollama>=0.2.0
langchain-core>=0.3.0
langgraph>=0.2.0
pydantic>=2.7.0
python-dotenv>=1.0.0
rdkit>=2023.9.5

# Testing
pytest>=8.2.0
pytest-mock>=3.14.0
